added print more qr code function

+++ lib/pages/Home/Detail/components/image/ImageCard.dart
@@ -4,6 +4,7 @@ import 'package:provider/provider.dart';
 import 'package:storage_management_mobile/DataObj/StorageItem.dart';
 import 'package:storage_management_mobile/States/ItemProvider.dart';
 import 'package:cached_network_image/cached_network_image.dart';
+import 'package:storage_management_mobile/pages/utils/ConfirmDialog.dart';
 
 class ImageCard extends StatelessWidget {
   final List<ImageObject> imageSrc;
@@ -22,9 +23,11 @@ class ImageCard extends StatelessWidget {
               ClipRRect(
                 borderRadius: BorderRadius.all(Radius.circular(0.0)),
                 child: CachedNetworkImage(
+                  key: Key("image-${i.id}"),
                   imageUrl: i.image,
                   height: MediaQuery.of(context).size.height,
                   fit: BoxFit.cover,
+                  placeholder: (context, url) => Text("image"),
                   progressIndicatorBuilder: (context, url, progress) => Center(
                     child: CircularProgressIndicator(
                       value: progress.progress,
@@ -38,7 +41,13 @@ class ImageCard extends StatelessWidget {
                   onPressed: () async {
                     ItemProvider pageState =
                         Provider.of(context, listen: false);
-                    await pageState.deleteImage(i.id);
+
+                    var confirm = await showConfirmDialog(context,
+                        title: "Delete image",
+                        content: "Do  you want to delete this?");
+                    if (confirm) {
+                      await pageState.deleteImage(i.id);
+                    }
                   },
                   icon: Icon(Icons.clear),
                 ),
